+++ telethon/client/__init__.py
@@ -23,4 +23,5 @@ from .account import AccountMethods
 from .auth import AuthMethods
 from .bots import BotMethods
 from .reactions import ReactionMethods
+from .payments import GiftMethods
 from .telegramclient import TelegramClient