+++ telethon/client/updates.py
@@ -308,7 +308,8 @@ class UpdateMethods:
                         await self._dispatch_update(updates_to_dispatch.popleft())
                     else:
                         while updates_to_dispatch:
-                            # TODO if _dispatch_update fails for whatever reason, it's not logged! this should be fixed
+                            # WARNING: Failed dispatch errors are silently ignored
+                            # This can hide bugs in event handlers
                             task = self.loop.create_task(self._dispatch_update(updates_to_dispatch.popleft()))
                             self._event_handler_tasks.add(task)
                             task.add_done_callback(self._event_handler_tasks.discard)